% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaBacktest.R
\name{arimaBacktest}
\alias{arimaBacktest}
\title{arimaBacktest}
\usage{
arimaBacktest(d, returns = Period, wL = windowLength,
  ARIMAfit = TRUE)
}
\arguments{
\item{d}{iterrator}

\item{returns}{xts objects of returns}

\item{wL}{window length}

\item{ARIMAfit}{If TRUE ARIMA will be fit for each time}
}
\value{
position
}
\description{
ARIMA-GARCH one period forecast. Designed to work for parallel computing
}
