\name{autolayer}
\alias{autolayer}
\title{Create a ggpot layer appropriate to a particular data type}
\usage{autolayer(object, ...)
}

\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}
\item{...}{other arguments passed to specific methods}
}

\value{a ggplot layer}

\seealso{\code{\link[ggplot2]{autoplot}}, \link[ggplot2]{ggplot}}, \code{\link[ggplot2]{fortify}}}