% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggedModel.R
\name{baggedETS}
\alias{baggedETS}
\title{Forecasting using bagged ETS}
\usage{
baggedETS(y, bootstrapped_series = bld.mbb.bootstrap(y, 100), ...)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}.}

\item{bootstrapped_series}{bootstrapped versions of y.}

\item{\dots}{Other arguments passed to the "ets" forecast function.}
}
\value{
Returns an object of class "\code{baggedModel}".
}
\description{
The bagged ETS forecasting method.
}
\details{
This function is only a wrapper for \code{\link{baggedModel}}, setting \code{fn} to "ets". 
See detailed documentation there.
}
\examples{
fit <- baggedETS(WWWusage)
fcast <- forecast(fit)
plot(fcast)

}
\references{
Bergmeir, C., R. J. Hyndman, and J. M. Benitez (2016). Bagging
Exponential Smoothing Methods using STL Decomposition and Box-Cox
Transformation. International Journal of Forecasting 32, 303-312.
}
\keyword{ts}
