% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cop_test_Par.R
\name{Cop_test_Par}
\alias{Cop_test_Par}
\title{Cop_test_Par}
\usage{
Cop_test_Par(d, returnsCop = Period, wL = windowLength,
  ARIMAfit = FALSE)
}
\arguments{
\item{d}{iterator}

\item{returnsCop}{xts of returns}

\item{wL}{window length}

\item{ARIMAfit}{If true, then an ARIMA will be fit for each iterations. Else a ARMA(2,1) is used}
}
\value{
Object of type "cGARCHfit". OBS a lot of non-fits objects are return as warnings or errors
}
\description{
Backtest using copula simulation. Function designed to be used for parallel
computing
}
