% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC_fore_Par.R
\name{DCC_fore_Par}
\alias{DCC_fore_Par}
\title{DCC_fore_Par}
\usage{
DCC_fore_Par(Period = Period, foreLength = foreLength,
  windowLength = windowLength, ARIMAfit = FALSE,
  outDir = "./output/", version = "v1", Name = NULL,
  tickers = tickers)
}
\arguments{
\item{Period}{xts object of log returns or other risk measures}

\item{foreLength}{length of forecast}

\item{windowLength}{length of window used for fitting}

\item{ARIMAfit}{If true, then an ARIMA will be fit for each iterations. Else a ARMA(2,1) is used}

\item{outDir}{Output dir}

\item{version}{Version  of dataset (if run multipe times and want to save each data)}

\item{Name}{Name of asset (colname)}

\item{tickers}{Name of columns}
}
\value{
Write to files position each day and its cumulative return
}
\description{
Simulate one day ahead forecast using DCC.
}
