% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC_test_Par.R
\name{DCC_test_Par}
\alias{DCC_test_Par}
\title{DCC_test_Par}
\usage{
DCC_test_Par(d, returnsDCC = Period, wL = windowLength,
  ARIMAfit = FALSE)
}
\arguments{
\item{d}{iterator}

\item{returnsDCC}{xts of returns}

\item{wL}{window length}

\item{ARIMAfit}{If true, then an ARIMA will be fit for each iterations. Else a ARMA(2,1) is used}
}
\value{
Object of type "DCCfit". OBS a lot of non-fits objects are return as warnings or errors
}
\description{
Backtest using DCC forecast. Function designed to be used for parallel
computing
}
