% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-package.R
\docType{package}
\name{forecast-package}
\alias{forecast-package}
\title{forecast: Forecasting Functions for Time Series and Linear Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods and tools for displaying and analysing univariate time series forecasts including exponential smoothing via state space models and automatic ARIMA modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/forecast/}
  \item \url{https://github.com/robjhyndman/forecast}
  \item Report bugs at \url{https://github.com/robjhyndman/forecast/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item George Athanasopoulos (\href{https://orcid.org/0000-0002-5389-2802}{ORCID})
  \item Christoph Bergmeir (\href{https://orcid.org/0000-0002-3665-9021}{ORCID})
  \item Gabriel Caceres (\href{https://orcid.org/0000-0002-2947-2023}{ORCID})
  \item Leanne Chhay
  \item Kirill Kuroptev
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
  \item Fotios Petropoulos (\href{https://orcid.org/0000-0003-3039-4955}{ORCID})
  \item Slava Razbash
  \item Earo Wang (\href{https://orcid.org/0000-0001-6448-5260}{ORCID})
  \item Farah Yasmeen (\href{https://orcid.org/0000-0002-1479-5401}{ORCID})
}

Other contributors:
\itemize{
  \item Federico Garza [contributor]
  \item Daniele Girolimetto [contributor]
  \item Ross Ihaka [contributor, copyright holder]
  \item R Core Team [contributor, copyright holder]
  \item Daniel Reid [contributor]
  \item David Shaub [contributor]
  \item Yuan Tang (\href{https://orcid.org/0000-0001-5243-233X}{ORCID}) [contributor]
  \item Xiaoqian Wang [contributor]
  \item Zhenyu Zhou [contributor]
}

}
\keyword{package}
