% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gas}
\alias{gas}
\title{Australian monthly gas production}
\format{
Time series data
}
\source{
Australian Bureau of Statistics.
}
\usage{
gas
}
\description{
Australian monthly gas production: 1956--1995.
}
\examples{
plot(gas)
seasonplot(gas)
tsdisplay(gas)

}
\keyword{datasets}
