\name{gold}
\alias{gold}
\title{Daily morning gold prices}
\description{Daily morning gold prices in US dollars. 1 January 1985 -- 31 March 1989.}
\usage{data(gold)}
\format{Time series data}
\source{Time Series Data Library. \url{http://data.is/TSDLdemo}}

\keyword{datasets}
\examples{tsdisplay(gold)
}
