% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{guide_train.level_colourbar}
\alias{guide_train.level_colourbar}
\title{Helper methods for guides}
\usage{
guide_train.level_colourbar(guide, scale)
}
\description{
Helper methods for guides
}
\keyword{internal}
