\name{guide_level_colourbar}
\alias{guide_level_colourbar}
\alias{guide_level_colorbar}
\alias{guide_level_legend}
\title{Create a ggplot layer appropriate to a particular data type}
\usage{guide_level_colourbar(...)
guide_level_colorbar(...)
guide_level_legend(...)
}

\arguments{
\item{...}{Arguments passed to \code{guide_colourbar} or \code{guide_legend}}
}

\description{
Replacement methods for \code{guide_colourbar} and \code{guide_legend}, for use with a level aesthetic.
}

\value{A guide object}

\seealso{\code{\link[ggplot2]{guide_colourbar}}, \code{\link[ggplot2]{guide_legend}}}
