% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newarima2.R
\name{is.constant}
\alias{is.constant}
\title{Is an object constant?}
\usage{
is.constant(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Returns true if the object's numerical values do not vary.
}
