\name{is.forecast}
\alias{is.forecast}
\alias{is.mforecast}
\alias{is.splineforecast}
\title{Is an object a particular forecast type?}
\usage{is.forecast(x)
is.mforecast(x)
is.splineforecast(x)
}

\arguments{
\item{x}{object to be tested}
}

\description{Returns true if the forecast object is of a particular type}