\name{is.ets}
\alias{is.ets}
\alias{is.acf}
\alias{is.Arima}
\alias{is.bats}
\alias{is.nnetar}
\alias{is.nnetarmodels}
\alias{is.stlm}
\title{Is an object a particular model type?}
\usage{is.ets(x)
is.acf(x)
is.Arima(x)
is.bats(x)
is.Arima(x)
is.bats(x)
is.ets(x)
is.nnetar(x)
is.nnetarmodels(x)
is.stlm(x)
}

\arguments{
\item{x}{object to be tested}
}

\description{Returns true if the model object is of a particular type}