% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkresiduals.R
\name{modeldf}
\alias{modeldf}
\title{Compute model degrees of freedom}
\usage{
modeldf(object, ...)
}
\arguments{
\item{object}{A time series model}

\item{...}{Other arguments currently ignored}
}
\description{
Compute model degrees of freedom
}
