% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{monthdays}
\alias{monthdays}
\title{Number of days in each season}
\usage{
monthdays(x)
}
\arguments{
\item{x}{time series}
}
\value{
Time series
}
\description{
Returns number of days in each month or quarter of the observed time period.
}
\details{
Useful for month length adjustments
}
\examples{

par(mfrow=c(2,1))
plot(ldeaths,xlab="Year",ylab="pounds",
    main="Monthly deaths from lung disease (UK)")
ldeaths.adj <- ldeaths/monthdays(ldeaths)*365.25/12
plot(ldeaths.adj,xlab="Year",ylab="pounds",
    main="Adjusted monthly deaths from lung disease (UK)")

}
\seealso{
\code{\link[forecast]{bizdays}}
}
\author{
Rob J Hyndman
}
\keyword{ts}
