% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{forecast}
\alias{accuracy}
\alias{autoplot}
\alias{autolayer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:accuracy]{accuracy()}}, \code{\link[generics:forecast]{forecast()}}}

  \item{ggplot2}{\code{\link[ggplot2:autolayer]{autolayer()}}, \code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}

