% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.nnetar}
\alias{residuals.nnetar}
\title{Residuals for various time series models}
\usage{
\method{residuals}{nnetar}(object, type = c("innovation", "response"),
  h = 1, ...)
}
\arguments{
\item{object}{An object containing a time series model of class \code{ar},
\code{Arima}, \code{bats}, \code{ets}, \code{fracdiff}, \code{nnetar} or
\code{stlm}.
If \code{object} is of class \code{forecast}, then the function will return
\code{object$residuals} if it exists, otherwise it returns the differences between
the observations and their fitted values.}

\item{type}{Type of residual.}

\item{h}{If \code{type='response'}, then the fitted values are computed for
\code{h}-step forecasts.}

\item{...}{Other arguments not used.}
}
\value{
A \code{ts} object of residuals. When \code{type="innovations"} and \code{lambda} is used, a matrix of time-series consisting of the residuals from each node in the network is returned.
}
\description{
Returns time series of residuals from a fitted model.
}
\details{
Innovation residuals correspond to the white noise process that drives the
evolution of the time series model. Response residuals are the difference
between the observations and the fitted values (equivalent to \code{h}-step
forecasts). For functions with no \code{h} argument, \code{h=1}. For
homoscedastic models, the innovation residuals and the response residuals
for \code{h=1} are identical. Regression residuals are available for
regression models with ARIMA errors, and are equal to the original data
minus the effect of the regression variables. If there are no regression
variables, the errors will be identical to the original series (possibly
adjusted to have zero mean).  \code{arima.errors} is a deprecated function
which is identical to \code{residuals.Arima(object, type="regression")}.
}
\examples{
fit <- Arima(lynx,order=c(4,0,0), lambda=0.5)

plot(residuals(fit))
plot(residuals(fit, type='response'))
}
\seealso{
\code{\link{fitted.Arima}}, \code{\link{checkresiduals}}.
}
