% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor}
\alias{taylor}
\title{Half-hourly electricity demand}
\format{Time series data}
\source{
James W Taylor
}
\usage{
taylor
}
\description{
Half-hourly electricity demand in England and Wales from Monday 5 June 2000
to Sunday 27 August 2000. Discussed in Taylor (2003), and kindly provided by
James W Taylor. Units: Megawatts
}
\examples{
plot(taylor)

}
\references{
Taylor, J.W. (2003) Short-term electricity demand forecasting
using double seasonal exponential smoothing. \emph{Journal of the
Operational Research Society}, \bold{54}, 799-805.
}
\keyword{datasets}
