% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_fore_Par.R
\name{VAR_fore_Par}
\alias{VAR_fore_Par}
\title{VAR_fore_Par}
\usage{
VAR_fore_Par(Period = Period, foreLength = foreLength,
  windowLength = windowLength, outDir = "./output/", version = "v1",
  Name = NULL, tickers = tickers)
}
\arguments{
\item{Period}{xts object of log returns or other risk measures}

\item{foreLength}{length of forecast}

\item{windowLength}{length of window used for fitting}

\item{outDir}{Output dir}

\item{version}{Version  of dataset (if run multipe times and want to save each data)}

\item{Name}{Name of asset (colname)}

\item{tickers}{Names of columns}
}
\value{
Write to files position each day and its cumulative return
}
\description{
Forecast oneday ahead VAR and backtest
}
