% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wineind}
\alias{wineind}
\title{Australian total wine sales}
\format{Time series data}
\source{
Time Series Data Library. \url{http://data.is/TSDLdemo}
}
\usage{
wineind
}
\description{
Australian total wine sales by wine makers in bottles <= 1 litre. Jan 1980
-- Aug 1994.
}
\examples{
tsdisplay(wineind)

}
\keyword{datasets}
