% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woolyrnq}
\alias{woolyrnq}
\title{Quarterly production of woollen yarn in Australia}
\format{
Time series data
}
\source{
Time Series Data Library. \url{https://pkg.yangzhuoranyang.com/tsdl/}
}
\usage{
woolyrnq
}
\description{
Quarterly production of woollen yarn in Australia: tonnes. Mar 1965 -- Sep
1994.
}
\examples{
tsdisplay(woolyrnq)

}
\keyword{datasets}
