% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.errors}
\alias{arima.errors}
\title{Residuals for various time series models}
\usage{
arima.errors(object)
}
\arguments{
\item{object}{An object containing a time series model of class \code{ar},
\code{Arima}, \code{bats}, \code{ets}, \code{fracdiff}, \code{nnetar} or
\code{stlm}. If object is of class \code{forecast},}
}
\value{
A \code{ts} object
}
\description{
Returns time series of residuals from a fitted model.
}
\details{
Innovation residuals correspond to the white noise process that drives the
evolution of the time series model. Response residuals are the difference
between the observations and the fitted values (equivalent to \code{h}-step
forecasts). For functions with no \code{h} argument, \code{h=1}. For
homoscedastic models, the innovation residuals and the response residuals
for \code{h=1} are identical. Regression residuals are available for
regression models with ARIMA errors, and are equal to the original data
minus the effect of the regression variables. If there are no regression
variables, the errors will be identical to the original series (possibly
adjusted to have zero mean).  \code{arima.errors} is a deprecated function
which is identical to \code{residuals.Arima(object, type="regression")}.
}
\seealso{
\code{\link{fitted.Arima}}, \code{\link{checkresiduals}}.
}
