% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.errors}
\alias{arima.errors}
\title{Errors from a regression model with ARIMA errors}
\usage{
arima.errors(object)
}
\arguments{
\item{object}{An object containing a time series model of class \code{Arima}.}
}
\value{
A \code{ts} object
}
\description{
Returns time series of the regression residuals from a fitted ARIMA model.
}
\details{
This is a deprecated function
which is identical to \code{\link{residuals.Arima}(object, type="regression")}
Regression residuals are equal to the original data
minus the effect of any regression variables. If there are no regression
variables, the errors will be identical to the original series (possibly
adjusted to have zero mean).
}
\seealso{
\code{\link{residuals.Arima}}.
}
\author{
Rob J Hyndman
}
\keyword{ts}
