% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{easter}
\alias{easter}
\title{Easter holidays in each season}
\usage{
easter(x, easter.mon = FALSE)
}
\arguments{
\item{x}{Monthly or quarterly time series}

\item{easter.mon}{If TRUE, the length of Easter holidays includes Easter
Monday.}
}
\value{
Time series
}
\description{
Returns a vector of 0's and 1's or fractional results if Easter spans March
and April in the observed time period. Easter is defined as the days from
Good Friday to Easter Sunday inclusively, plus optionally Easter Monday if
\code{easter.mon=TRUE}.
}
\details{
Useful for adjusting calendar effects.
}
\examples{

  easter(wineind, easter.mon = TRUE)

}
\author{
Earo Wang
}
\keyword{ts}
