% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\docType{package}
\name{forecast-package}
\alias{forecast-package}
\title{Forecasting Functions for Time Series and Linear Models}
\description{
Methods and tools for displaying and analysing univariate time series
forecasts including exponential smoothing via state space models and
automatic ARIMA modelling.
}
\details{
\tabular{ll}{ Package: \tab forecast\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\author{
Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{package}
