% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponse.R
\name{getResponse}
\alias{getResponse}
\alias{getResponse.default}
\alias{getResponse.lm}
\alias{getResponse.Arima}
\alias{getResponse.fracdiff}
\alias{getResponse.ar}
\alias{getResponse.tbats}
\alias{getResponse.bats}
\alias{getResponse.mforecast}
\alias{getResponse.baggedModel}
\title{Get response variable from time series model.}
\usage{
getResponse(object, ...)

\method{getResponse}{default}(object, ...)

\method{getResponse}{lm}(object, ...)

\method{getResponse}{Arima}(object, ...)

\method{getResponse}{fracdiff}(object, ...)

\method{getResponse}{ar}(object, ...)

\method{getResponse}{tbats}(object, ...)

\method{getResponse}{bats}(object, ...)

\method{getResponse}{mforecast}(object, ...)

\method{getResponse}{baggedModel}(object, ...)
}
\arguments{
\item{object}{a time series model or forecast object.}

\item{...}{Additional arguments that are ignored.}
}
\value{
A numerical vector or a time series object of class \code{ts}.
}
\description{
\code{getResponse} is a generic function for extracting the historical data
from a time series model (including \code{Arima}, \code{ets}, \code{ar},
\code{fracdiff}), a linear model of class \code{lm}, or a forecast object.
The function invokes particular \emph{methods} which depend on the class of
the first argument.
}
\author{
Rob J Hyndman
}
\keyword{ts}
