% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gold}
\alias{gold}
\title{Daily morning gold prices}
\format{
Time series data
}
\usage{
gold
}
\description{
Daily morning gold prices in US dollars. 1 January 1985 -- 31 March 1989.
}
\examples{
tsdisplay(gold)

}
\keyword{datasets}
