% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{guide_train.level_guide}
\alias{guide_train.level_guide}
\alias{guide_geom.guide_level}
\title{Helper methods for guides}
\usage{
guide_train.level_guide(guide, scale, aesthetic)

guide_geom.guide_level(guide, layers, default_mapping)
}
\description{
Helper methods for guides
}
\keyword{internal}
