% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{guide_level}
\alias{guide_level}
\title{Level shade bar guide}
\usage{
guide_level(title = waiver(), max_discrete = 5, ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[=waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[=labs]{labs()}} is used for the title.}

\item{...}{ignored.}
}
\description{
The level guide shows the colour from the forecast intervals which is blended with the series colour.
}
