% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R, R/arima.R, R/baggedModel.R, R/bats.R,
%   R/ets.R, R/modelAR.R, R/mstl.R, R/nnetar.R
\name{is.acf}
\alias{is.acf}
\alias{is.Arima}
\alias{is.baggedModel}
\alias{is.bats}
\alias{is.ets}
\alias{is.modelAR}
\alias{is.stlm}
\alias{is.nnetar}
\alias{is.nnetarmodels}
\title{Is an object a particular model type?}
\usage{
is.acf(x)

is.Arima(x)

is.baggedModel(x)

is.bats(x)

is.ets(x)

is.modelAR(x)

is.stlm(x)

is.nnetar(x)

is.nnetarmodels(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Returns true if the model object is of a particular type
}
