% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{na.interp}
\alias{na.interp}
\title{Interpolate missing values in a time series}
\usage{
na.interp(x, lambda = NULL)
}
\arguments{
\item{x}{time series}

\item{lambda}{Box-Cox decomposition parameter. If \code{NULL}, no transformation
is used. If \code{lambda="auto"}, a transformation is automatically selected. If
lambda takes a numerical value, it is used as the parameter of the Box-Cox transformation.}
}
\value{
Time series
}
\description{
Uses linear interpolation for non-seasonal series and a periodic stl
decomposition with seasonal series to replace missing values.
}
\details{
A more general and flexible approach is available using \code{na.approx} in
the \code{zoo} package.
}
\examples{

data(gold)
plot(na.interp(gold))

}
\seealso{
\code{\link[forecast]{tsoutliers}}
}
\author{
Rob J Hyndman
}
\keyword{ts}
