\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
  \item{ggplot2}{\code{autoplot}}
}}

