% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{scale_level}
\alias{scale_level}
\alias{scale_level_hue}
\alias{scale_level_brewer}
\alias{scale_level_distiller}
\alias{scale_level_gradient2}
\alias{scale_level_gradientn}
\alias{scale_level_grey}
\alias{scale_level_identity}
\alias{scale_level_manual}
\alias{scale_level_gradient}
\alias{scale_level_continuous}
\alias{scale_level_discrete}
\title{Edge colour scales}
\usage{
scale_level_hue(..., h = c(0, 360) + 15, c = 100, l = 65, h.start = 0,
  direction = 1, na.value = "grey50")

scale_level_brewer(..., type = "seq", palette = 1, direction = 1)

scale_level_distiller(..., type = "seq", palette = 1, direction = -1,
  values = NULL, space = "Lab", na.value = "grey50",
  guide = "level_colourbar")

scale_level_gradient2(..., low = muted("red"), mid = "white",
  high = muted("blue"), midpoint = 0, space = "Lab",
  na.value = "grey50", guide = "level_colourbar")

scale_level_gradientn(..., colours, values = NULL, space = "Lab",
  na.value = "grey50", guide = "level_colourbar", colors)

scale_level_grey(..., start = 0.2, end = 0.8, na.value = "red")

scale_level_identity(..., guide = "none")

scale_level_manual(..., values)

scale_level_gradient(..., low = "#888888", high = "#BBBBBB",
  space = "Lab", na.value = NA, guide = "level_colourbar")

scale_level_continuous(..., low = "#888888", high = "#BBBBBB",
  space = "Lab", na.value = NA, guide = "level_colourbar")

scale_level_discrete(..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, na.value = "grey50")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}

\item{h}{range of hues to use, in [0, 360]}

\item{c}{chroma (intensity of colour), maximum value varies depending on
combination of hue and luminance.}

\item{l}{luminance (lightness), in [0, 100]}

\item{h.start}{hue to start at}

\item{direction}{direction to travel around the colour wheel,
1 = clockwise, -1 = counter-clockwise}

\item{na.value}{Colour to use for missing values}

\item{type}{One of seq (sequential), div (diverging) or qual (qualitative)}

\item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link{rescale}} for a convience function
to map an arbitrary range to between 0 and 1.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{mid}{colour for mid point}

\item{midpoint}{The midpoint (in data value) of the diverging scale.
Defaults to 0.}

\item{colours, colors}{Vector of colours to use for n-colour gradient.}

\item{start}{gray value at low end of palette}

\item{end}{gray value at high end of palette}
}
\value{
A ggproto object inheriting from `Scale`
}
\description{
This set of scales defines new colour scales for edge geoms equivalent to the
ones already defined by ggplot2. The parameters are equivalent to the ones
from ggplot2 so there is nothing new under the sun. The different geoms will
know whether to use edge scales or the standard scales so it is not necessary
to write `level` in the call to the geom - just use `colour`.
}
