% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{scale_level}
\alias{scale_level}
\alias{scale_level_gradient}
\alias{scale_level_continuous}
\title{Level colour scales}
\usage{
scale_level_gradient(..., low = "#888888", high = "#BBBBBB",
  space = "Lab", na.value = NA, guide = "level_colourbar")

scale_level_continuous(..., low = "#888888", high = "#BBBBBB",
  space = "Lab", na.value = NA, guide = "level_colourbar")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{continuous_scale}}
to control name, limits, breaks, labels and so forth.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}
}
\value{
A ggproto object inheriting from `Scale`
}
\description{
This set of scales defines new scales for level geoms equivalent to the
ones already defined by ggplot2. This allows the shade of confidence intervals
to work with the legend output.
}
