% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasadj.R
\name{seasadj}
\alias{seasadj}
\alias{seasadj.stl}
\alias{seasadj.mstl}
\alias{seasadj.decomposed.ts}
\alias{seasadj.tbats}
\alias{seasadj.seas}
\title{Seasonal adjustment}
\usage{
seasadj(object, ...)

\method{seasadj}{stl}(object, ...)

\method{seasadj}{mstl}(object, ...)

\method{seasadj}{decomposed.ts}(object, ...)

\method{seasadj}{tbats}(object, ...)

\method{seasadj}{seas}(object, ...)
}
\arguments{
\item{object}{Object created by \code{\link[stats:decompose]{stats::decompose()}}, \code{\link[stats:stl]{stats::stl()}} or
\code{\link[=tbats]{tbats()}}.}

\item{...}{Other arguments not currently used.}
}
\value{
Univariate time series.
}
\description{
Returns seasonally adjusted data constructed by removing the seasonal
component.
}
\examples{
plot(AirPassengers)
lines(seasadj(decompose(AirPassengers, "multiplicative")), col = 4)

}
\seealso{
\code{\link[stats:stl]{stats::stl()}}, \code{\link[stats:decompose]{stats::decompose()}}, \code{\link[=tbats]{tbats()}}.
}
\author{
Rob J Hyndman
}
\keyword{ts}
