\name{taylor}
\alias{taylor}
\title{Half-hourly electricity demand}
\description{Half-hourly electricity demand in England and Wales from Monday 5 June 2000 to Sunday 27 August 2000. Discussed in Taylor (2003), and kindly provided by James W Taylor.}
\usage{taylor}
\format{Time series data}
\source{James W Taylor}
\references{Taylor, J.W. (2003) Short-term electricity demand forecasting using double seasonal exponential smoothing. \emph{Journal of the Operational Reseach Society}, \bold{54}, 799-805.
}


\keyword{datasets}
\examples{plot(taylor)
}
