\name{wineind}
\alias{wineind}
\title{Australian total wine sales}
\description{Australian total wine sales by wine makers in bottles <= 1 litre. Jan 1980 -- Aug 1994.}
\usage{wineind}
\format{Time series data}
\source{Time Series Data Library. \url{http://data.is/TSDLdemo}}

\keyword{datasets}
\examples{tsdisplay(wineind)
}
