\name{woolyrnq}
\alias{woolyrnq}
\title{Quarterly production of woollen yarn in Australia}
\description{Quarterly production of woollen yarn in Australia: tonnes. Mar 1965 -- Sep 1994.}
\usage{woolyrnq}
\format{Time series data}
\source{Time Series Data Library. \url{http://data.is/TSDLdemo}}

\keyword{datasets}
\examples{tsdisplay(woolyrnq)
}
